requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Deprecated==1.2.14
distro==1.9.0
frozenlist==1.4.1
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
line-bot-sdk==3.7.0
multidict==6.0.4
openai==1.8.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4